+++ src/commands/hardis/source/deploy.ts
@@ -14,6 +14,8 @@ Additional to the base command wrapper: If using **--checkonly**, add options **
 You can also have deployment results as pull request comments, on:
 
 - Gitlab (see [Gitlab integration configuration](https://sfdx-hardis.cloudity.com/salesforce-ci-cd-setup-integration-gitlab/))
+- Azure DevOps (see [Azure integration configuration](https://sfdx-hardis.cloudity.com/salesforce-ci-cd-setup-integration-azure/))
+
 
 [![Assisted solving of Salesforce deployments errors](https://github.com/hardisgroupcom/sfdx-hardis/raw/main/docs/assets/images/article-deployment-errors.jpg)](https://nicolas.vuillamy.fr/assisted-solving-of-salesforce-deployments-errors-47f3666a9ed0)
 
@@ -122,9 +124,10 @@ You can also have deployment results as pull request comments, on:
     // Check org coverage if requested
     if (this.flags.checkcoverage && result.stdout) {
       const orgCoveragePercent = await extractOrgCoverageFromLog(result.stdout + result.stderr || "");
+      const checkOnly = this.flags.checkonly || false;
       if (orgCoveragePercent) {
         try {
-          await checkDeploymentOrgCoverage(orgCoveragePercent);
+          await checkDeploymentOrgCoverage(orgCoveragePercent, { check: checkOnly });
         } catch (errCoverage) {
           await GitProvider.managePostPullRequestComment();
           throw errCoverage;