+++ src/commands/hardis/work/new.ts
@@ -167,33 +167,35 @@ Under the hood, it can:
     // Get allowed work org types from config if possible
     const allowedOrgTypes = config?.allowedOrgTypes || [];
     let selectedOrgType = allowedOrgTypes.length == 1 ? allowedOrgTypes[0] : null;
-    // If necessary, Prompt if you want to use a scratch org or a tracked sandbox org
-    if (selectedOrgType == null) {
-      const orgTypeResponse = await prompts({
-        type: "select",
-        name: "value",
-        message: c.cyanBright(`Do you want to use a scratch org or a tracked sandbox org ?`),
-        initial: 0,
-        choices: [
-          {
-            title: "Scratch org",
-            value: "scratch",
-            description: "Scratch orgs are configured on my project so I want to create or reuse one",
-          },
-          {
-            title: "Sandbox org with source tracking",
-            value: "sandbox",
-            description: "Release manager told me that I can work on Sandboxes on my project so let's use fresh dedicated one",
-          },
-          {
-            title: "I'm hardcore, I don't need an org !",
-            value: "noOrg",
-            description: "You just want to play with XML and sfdx-hardis configuration, and you know what you are doing !",
-          },
-        ],
+    // If necessary, Prompt if you want to use a scratch org or a tracked sandbox org, or no org
+    const orgTypeChoices = [];
+    if (allowedOrgTypes.includes("sandbox") || allowedOrgTypes.length === 0) {
+      orgTypeChoices.push({
+        title: "Sandbox org with source tracking",
+        value: "sandbox",
+        description: "Release manager told me that I can work on Sandboxes on my project so let's use fresh dedicated one",
       });
-      selectedOrgType = orgTypeResponse.value;
     }
+    if (allowedOrgTypes.includes("scratch") || allowedOrgTypes.length === 0) {
+      orgTypeChoices.push({
+        title: "Scratch org",
+        value: "scratch",
+        description: "Scratch orgs are configured on my project so I want to create or reuse one",
+      });
+    }
+    orgTypeChoices.push({
+      title: "I'm hardcore, I don't need an org !",
+      value: "noOrg",
+      description: "You just want to play with XML and sfdx-hardis configuration, and you know what you are doing !",
+    });
+    const orgTypeResponse = await prompts({
+      type: "select",
+      name: "value",
+      message: c.cyanBright(`Do you want to use a scratch org or a tracked sandbox org ?`),
+      initial: 0,
+      choices: orgTypeChoices,
+    });
+    selectedOrgType = orgTypeResponse.value;
 
     // Select or create org that user will work in
     if (selectedOrgType === "scratch") {