+++ src/common/gitProvider/gitProviderRoot.ts
@@ -13,7 +13,19 @@ export abstract class GitProviderRoot {
 
   public async postPullRequestMessage(prMessage: PullRequestMessageRequest): Promise<PullRequestMessageResult> {
     uxLog(this, c.yellow("Method postPullRequestMessage is not yet implemented on " + this.getLabel() + " to post " + JSON.stringify(prMessage)));
-    return null;
+    return { posted: false, providerResult: { error: "Not implemented in sfdx-hardis" } };
+  }
+
+  // Do not make crash the whole process in case there is an issue with integration
+  public async tryPostPullRequestMessage(prMessage: PullRequestMessageRequest): Promise<PullRequestMessageResult> {
+    let prResult: PullRequestMessageResult = null;
+    try {
+      prResult = await this.postPullRequestMessage(prMessage);
+    } catch (e) {
+      uxLog(this, c.yellow(`[GitProvider] Error while trying to post pull request message.\n${e.message}\n${e.stack}`));
+      prResult = { posted: false, providerResult: { error: e } };
+    }
+    return prResult;
   }
 
   protected async getParentMergeRequestId(): Promise<number> {